requirements.txt
sasl
thrift
thrift-sasl
PyHive
psycopg2-binary
